requirements.txt
numpy>=1.20
Pillow>=7.1.2
--find-links https://download.pytorch.org/whl/cu111/torch_stable.html
torchvision==0.10.0
tqdm>=4.41.0
tensorboard>=2.4.1
pycocotools>=2.0.4